requirements.txt
pandas==2.2.3
pydantic==2.11.3
pydantic[email]
reportlab==4.3.1
python-dotenv==1.1.0